src/faster_rcnn/requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
bitsandbytes==0.47.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
datasets==4.1.1
decorator==5.2.1
dill==0.4.0
docker-pycreds==0.4.0
docstring_parser==0.17.0
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.31.2
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.0
jsonargparse==4.41.0
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.1.post0
lightning-utilities==0.14.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.3.2
peft==0.15.2
pillow==11.2.1
pip-autoremove==0.10.0
platformdirs==4.3.8
proglog==0.1.12
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==21.0.0
pycocotools==2.0.8
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
qudida==0.0.4
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==80.7.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
stringzilla==3.12.5
sympy==1.14.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.2
torch==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
transformers==4.53.0
triton==3.3.0
typeshed_client==2.8.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.11
xxhash==3.5.0
yarl==1.20.0
